+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fixed ?order (if scalar it does not add "ASC")
 * Added to travis-ci.org
 * ConfigError: normal message
+* #55: Extended ?table format
 
 ## 2.2.6 (12.02.2016)